% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{cleanCols}
\alias{cleanCols}
\title{Function to clean sensata data}
\usage{
cleanCols(
  df,
  dictionary,
  colsToKeep = c("id", "surveyName", "surveyId", "totalTimeMin",
    "geolocation.coordinates", "fingerprint", "browserReport.ip", "sensataId",
    "createdAt"),
  removeParams = FALSE,
  removeScreens = TRUE
)
}
\arguments{
\item{df}{data downloaded from Mongo and loaded to R.}

\item{dictionary}{dictionary created with dictGenerator.R}

\item{colsToKeep}{vector of names of metadata columns to keep. Defaults to: id, surveyName, surveyId, totalTimeMin, geolocation.coordinates, fingerprint, browserReport.ip and, createdAt. All questions are always included.}

\item{removeParams}{logical, if TRUE params from mongo are removed.}

\item{removeScreens}{logical, if TRUE  screens are removed}
}
\value{
Dataframe with renamed fewer metadata and renamed selected answers as the question identifier.
}
\description{
This function eliminates the unnecesary metadata columns and renames the columns from mongo to the identifier, according to the dictionary.
}
\examples{
bogData <- bogData \%>\% cleanCols(dictionary = bogDic)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
