% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Function to clean sensata data}
\usage{
cleanData(
  data,
  dictionary,
  keep = c("id", "surveyName", "surveyId", "totalTimeMin", "geolocation.coordinates",
    "fingerprint", "browserReport.ip", "sensataId", "createdAt"),
  params = T,
  skipVal = "Saltar pregunta"
)
}
\arguments{
\item{data}{data downloaded from Mongo and loaded to R.}

\item{dictionary}{dictionary created with dictGenerator.R}

\item{keep}{vector of names of metadata columns to keep. Defaults to: id, surveyName, surveyId, totalTimeMin, geolocation.coordinates, fingerprint, browserReport.ip and, createdAt. All questions are always included.}

\item{params}{logical, if F params from mongo are erased.}

\item{skipVal}{Text that should have all the skipped questions. defaults to "Saltar pregunta"}
}
\value{
Dataframe dictionary and exported to .csv and .Rdata useful for data cleaning
}
\description{
This function eliminates the unnecesary metadata columns and renames the columns from mongo to the identifier, according to the dictionary.
}
\examples{
cleanData(dictionary = bogExpDic,skipVal = "Probando skipVal")
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{diccionario}
\keyword{dictionary}
\keyword{sensata}
