% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint2Tasks.R
\name{conjoint2Tasks}
\alias{conjoint2Tasks}
\title{Function to create conjoint data by rounds}
\usage{
conjoint2Tasks(df, nRounds, nInd)
}
\arguments{
\item{df}{conjoint Dataframe already scrubbed and cleaned of extra characters}

\item{nRounds}{number of rounds/task of the conjoint}

\item{nInd}{number of individuals}
}
\value{
dataframe with rows as task done
}
\description{
This function eliminates extra characters, usually used for conjoint. For now it only eliminates the characters from the conjoint options field
}
\examples{
CJData <- conjoint2Tasks(CJData, nRounds = 3, nInd = 42284)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{conjoint}
\keyword{prepare}
\keyword{sensata}
