% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delExtraCar}
\alias{delExtraCar}
\title{Function to eliminate extra characters for conjoint}
\usage{
delExtraCar(x)
}
\arguments{
\item{x}{conjoint options column (or character vector)}
}
\value{
character vector without extra characters
}
\description{
This function eliminates extra characters, usually used for conjoint. For now it only eliminates the characters from the conjoint options field
}
\examples{
delExtraCar(cjData$conjoint.0.options)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{characters}
\keyword{clean}
\keyword{conjoint}
\keyword{sensata}
