% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccFunction.R
\name{diccFunction}
\alias{diccFunction}
\title{Function to create sensata dictionary}
\usage{
diccFunction(dat, exclude = NULL, f = NULL)
}
\arguments{
\item{dat}{datafile downloaded from Mongo}

\item{exclude}{question (ex. q2) to be excluded from the dictionary (ex because is an exit screen)}

\item{f}{file path to save the dictionaries}
}
\value{
Dataframe dictionary and exported to .csv and .Rdata useful for data cleaning
}
\description{
This function allows you to create a dictionary from Sensata's data exported from Mongo.
}
\examples{
diccFunction("IND_data.csv", q2)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{diccionario}
\keyword{dictionary}
\keyword{sensata}
