% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dividehun_fun}
\alias{dividehun_fun}
\title{Divide by a hundred (usually to change percentage to proportion)}
\usage{
dividehun_fun(x)
}
\arguments{
\item{x}{vector to divide by a hundred (ex. from percentage to proportion)}
}
\value{
vector to be divided by a hundred
}
\description{
This function divides any vector by a hundred, we use it regularly to convert percentage to proportion
}
\examples{
dividethou_fun(data$percentage)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{divide}
\keyword{hundreds}
\keyword{percentage}
\keyword{proportion}
\keyword{sensata}
