% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dividethou_fun}
\alias{dividethou_fun}
\title{Divide by a thousand (usually to convert milliseconds to seconds)}
\usage{
dividethou_fun(x)
}
\arguments{
\item{x}{vector to divide by a thousand (ex. from milliseconds to seconds)}
}
\value{
vector to be divided by a thousand
}
\description{
This function divides any vector by a thousand, we use it regularly to convert seconds to milliseconds
}
\examples{
dividethou_fun(data$q1time)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{divide}
\keyword{milliseconds}
\keyword{seconds}
\keyword{segundos}
\keyword{sensata}
\keyword{thousands}
