% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetaData.R
\name{extractMetaData}
\alias{extractMetaData}
\title{Extract sensata metadata}
\usage{
extractMetaData(
  df = intData,
  dict = dict,
  stage = NULL,
  source = NULL,
  medium = NULL
)
}
\arguments{
\item{df}{Interim dataset after scrubRows.r}

\item{dict}{Dictionary created using dictGenerator.R}

\item{stage}{pilot or live collection for params.utm_campaign}

\item{source}{wa, fb, email, etc for params.utm_source}

\item{medium}{cpc, referral, etc for params.utm_medium}
}
\value{
1-row metadata for columns A to H (except lang)
}
\description{
This function creates table to fill out metadata on https://docs.google.com/spreadsheets/d/1CdSqvX3Hhayk_NGvr9eqDeAlA8XG9BX_-Exm5K3RKlI/edit#gid=2106088199
}
\examples{
extractMetaData(data$q1time)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{extract}
\keyword{metadata}
\keyword{sensataDataProg}
