% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorSensata.R
\name{factorSensata}
\alias{factorSensata}
\title{Function to create factors from sensata data}
\usage{
factorSensata(
  data,
  dictionary,
  questionPrefix = "q_",
  skipVal = "Saltar pregunta"
)
}
\arguments{
\item{dictionary}{dictionary created using dictGenerator.R in expanded form}

\item{questionPrefix}{Character that identifies questions}

\item{skipVal}{value that represents the skipped questions. By default "Saltar pregunta"}

\item{dat}{data downloaded from Mongo, cleaned with cleanData.R and scrubbed with scrubData.R}
}
\value{
Dataframe dictionary and exported to .csv and .Rdata useful for data cleaning
}
\description{
This function allows you to create factors with the information from the dictionary to the variables in the microdata.
DO NOT USE for NPS questions that have range different than 0-10.
If there are some multiple choice questions, it will show warnings, because it will add NA's to people that did not select all possible options
}
\examples{
diccFunction("IND_data.csv", q2)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{dictionary}
\keyword{factors}
\keyword{microdata}
\keyword{sensata}
