% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFactors.R
\name{makeFactors}
\alias{makeFactors}
\title{Function to create factors from sensata data}
\usage{
makeFactors(
  df,
  dictionary,
  questionPrefix = "q_",
  skipQuestionString = "Saltar pregunta",
  multChoiceText = c(1, 0),
  dummyMultiChoice = T
)
}
\arguments{
\item{df}{data downloaded from Mongo, cleaned with cleanData.R and scrubbed with scrubData.R}

\item{dictionary}{dictionary created using dictGenerator.R in expanded form without screens}

\item{questionPrefix}{Character that identifies questions}

\item{skipQuestionString}{value that represents the skipped questions. By default "Saltar pregunta"}

\item{multChoiceText}{A two element vector that determines text of multiple choice questions. By defaul c(1,0). Commonly used ones are c("Yes", "No") or c("Sí", "No"). ORDER MATTERS!}

\item{dummyMultiChoice}{logic, if TRUE (default), it will create a dummy column for each option on multiple choice questions. Usually turned off when making factors after translation.}
}
\value{
Dataframe with questions in format that is analysis friendly.
}
\description{
This function allows you to create factors with the information from the dictionary to the variables in the microdata.
DO NOT USE for NPSor slider questions that have range different than 0-10.
}
\examples{
coronaTestData <-makeFactors(df = bogData3, dict = bogExpDic)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{dictionary}
\keyword{factors}
\keyword{microdata}
\keyword{sensata}
