% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctions.R
\name{mis_fun}
\alias{mis_fun}
\title{Percentage of missing values}
\usage{
mis_fun(x, mis = "NS-NR", nques)
}
\arguments{
\item{x}{vector to count missings on}

\item{mis}{character that identifies missing values.}

\item{nques}{number of questions on the instrument}
}
\value{
percentage of that vector that is identical to missing value
}
\description{
This function counts how many of a vector are missing, and divides them by the number of questions
}
\examples{
dividethou_fun(data$q1time)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{missing}
\keyword{sensata}
\keyword{skip}
