% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubData.R
\name{scrubData}
\alias{scrubData}
\title{Function to scrub sensata data}
\usage{
scrubData(data, dup = T, ageVar, ageVal, timeMin = 2.5, geoloc = F)
}
\arguments{
\item{data}{data downloaded from Mongo and cleaned with cleanData.R}

\item{dup}{logical, if FALSE it does not scrub duplicate data}

\item{ageVar}{name of variable of age variable, if empty then it will not scrub by age.}

\item{ageVal}{value of age that should be excluded}

\item{timeMin}{minimum amount of minutes that the survey should have. Default 2.5 mins}

\item{geoloc}{logical, if TRUE it will scrub surveys that have no geolocation.}
}
\value{
Dataframe dictionary and exported to .csv and .Rdata useful for data cleaning
}
\description{
This function eliminates cases because they are duplicated, or minors, or other reasons.
}
\examples{
scrubData("IND_data.csv", q2)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{diccionario}
\keyword{dictionary}
\keyword{sensata}
