% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubPeaks.R
\name{scrubPeaks}
\alias{scrubPeaks}
\title{Function to scrub atypical peaks to control manipulation attempts using medium param}
\usage{
scrubPeaks(df, iniPico = NULL, finPico = NULL, mediumToScrub = NULL)
}
\arguments{
\item{df}{microdata after cleanCols and scrubRows}

\item{iniPico}{datetime beginning of peak ex. "2021-08-06 07:00:00", default: NULL if you need to drop all surveys from one medium.}

\item{finPico}{datetime beginning of peak ex. "2021-08-09 07:00:00", default: NULL if you need to drop all surveys from one medium.}

\item{mediumToScrub}{medium value to scrub ex. "referral", default: NULL if you need to drop all surveys from a date range}
}
\value{
microdata without rows from atypical peaks and attributes saving number of individuals
}
\description{
This function creates the microdata after supressing atypical peaks
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{CCV21-1}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
