% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrubRows.R
\name{scrubRows}
\alias{scrubRows}
\title{Function to scrub sensata data}
\usage{
scrubRows(
  df,
  removeDupes = T,
  timeMin = 2.5,
  geoloc = F,
  ageVar = NULL,
  ageVal = NULL,
  testParamName = NULL,
  completeVars = NULL
)
}
\arguments{
\item{df}{data downloaded from Mongo and cleaned with cleanData.R}

\item{removeDupes}{logical, if TRUE it scrubs removeDupeslicate data}

\item{timeMin}{minimum amount of minutes that the survey should have. Default 2.5 mins. If no scrubbing by time is required make it 0.}

\item{geoloc}{logical, if TRUE it will scrub surveys that have no geolocation.}

\item{ageVar}{name of variable of age variable, if empty then it will not scrub by age.}

\item{ageVal}{value of age that should be excluded}

\item{testParamName}{character object of name of test param, usually test (the full column is called params.test)}

\item{completeVars}{character vector of variables that have to be complete. It erases individuals that did not answer ALL of them.}
}
\value{
Dataframe with the cases scrubbed, and the attributes with the number of cases left after each step for the report: oriNum, removeDupesNum, ageNum, timeNum, geoNum, and finNum
}
\description{
This function eliminates cases because they are duplicates, or minors, or due to duration of the survey, or because it doesn't have geolocation coordinates.
It also creates attributes with the number of people left after each scrubbing step.
}
\examples{
bogData1 <- bogData1 \%>\% scrubData(removeDupes = F, ageVar = "EVCS2", ageVal = "Menos de 18 años", geoloc = T)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{data-scrubbing}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
