% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensataDataProg-package.R
\docType{package}
\name{sensataDataProg}
\alias{sensataDataProg}
\title{sensataDataProg: This package has the functions needed to clean and manipulate data created with the Sensata platform.}
\description{
Este paquete tiene las funciones necesarias para limpiar y manipular los datos de la plataforma Sensata.
}
\section{sensataDataProg basic functions}{

El paquete provee funciones para los pasos más importantes en la depuración de los datos:
\itemize{
\item dictGenerator:  crear un diccionario
\item cleanCols: limpiar las columnas
\item scrubRows: depurar los datos
\item makeFactors: crear factores para el análisis
\item selectCols: dejar las columnas finales para el cliente
}
}

\section{Additional functions}{

Además incluye funciones para otros procesamientos de datos que solemos hacer:
\itemize{
\item scrubPeaks: eliminar picos atípicos
\item translateFactors: traducir datos de un idioma a otro
}
}

