% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensataExample}
\alias{sensataExample}
\title{Raw sensata data example.}
\format{
A data frame with 590 rows and 609 variables:
\describe{
\item{newresponses}{Information on responses}
\item{othercols}{other metadata}
}
}
\source{
hgsLiveRaw with a fake params.test
}
\usage{
sensataExample
}
\description{
A dataset of survey 590 surveys responses downloaded from mongo, HGS survey
sensataExample
}
\keyword{datasets}
