% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Function to clean sensata data}
\usage{
cleanData(
  data,
  dictionary,
  keep = c("id", "surveyName", "surveyId", "totalTimeMin", "geolocation.coordinates",
    "fingerprint", "browserReport.ip", "sensataId", "createdAt"),
  removeParams = F,
  skipVal = "Saltar pregunta",
  questionPrefix = "q_"
)
}
\arguments{
\item{data}{data downloaded from Mongo and loaded to R.}

\item{dictionary}{dictionary created with dictGenerator.R}

\item{keep}{vector of names of metadata columns to keep. Defaults to: id, surveyName, surveyId, totalTimeMin, geolocation.coordinates, fingerprint, browserReport.ip and, createdAt. All questions are always included.}

\item{removeParams}{logical, if T params from mongo are removed}

\item{skipVal}{Text that should have all the skipped questions. defaults to "Saltar pregunta"}

\item{questionPrefix}{Character that identifies questions.}
}
\value{
Dataframe with only the necessary metadata columns, usually called as interim data.
}
\description{
This function eliminates the unnecesary metadata columns and renames the columns from mongo to the identifier, according to the dictionary.
}
\examples{
bogData <- bogData \%>\% cleanData(dictionary = bogDic)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
