% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictGenerator.R
\name{dictGenerator}
\alias{dictGenerator}
\title{Function to create sensata dictionary}
\usage{
dictGenerator(
  df,
  cols = c("qid", "identifier", "question", "type", "options", "numberOfOptions",
    "isOrdered", "isSorting", "maxResponses", "altOption"),
  expandOptions = TRUE,
  questionPrefix = "q_",
  forceOrdered = NULL,
  responseType = "newResponses"
)
}
\arguments{
\item{df}{data downloaded from Mongo and loaded to R}

\item{cols}{columns to create on the dictionary.}

\item{expandOptions}{logical. If TRUE, the dictionary will have one row per answer option. if FALSE, it will be 1 row per question.}

\item{questionPrefix}{character that identifies questions}

\item{forceOrdered}{vector of ordered questions that are from a non-ordered types}

\item{responseType}{which object version should the function use, one of "newResponses" or "structuredResponses"}
}
\value{
Dataframe dictionary useful for data cleaning
}
\description{
This function allows you to create a dictionary from Sensata's data exported from Mongo.
}
\examples{
dictGenerator(data = demoData, expandOptions = F)
}
\author{
Camilo Delvasto \email{camilo@sensata.io} & Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{diccionario}
\keyword{dictionary}
\keyword{sensata}
