% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictGenerator.R
\name{dictGenerator}
\alias{dictGenerator}
\title{Function to create sensata dictionary}
\usage{
dictGenerator(
  data,
  params = c("identifier", "question", "type", "options"),
  expandOptions = T
)
}
\arguments{
\item{data}{data downloaded from Mongo and loaded to R}

\item{params}{columns to create on the dictionary.}

\item{expandOptions}{logical. If FALSE, the dictionary will have one row per question.}
}
\value{
Dataframe dictionary useful for data cleaning
}
\description{
This function allows you to create a dictionary from Sensata's data exported from Mongo.
}
\examples{
dictGenerator(data = demoData, params = params, expandOptions = F)
}
\author{
Camilo Delvasto \email{camilo@sensata.io} & Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{diccionario}
\keyword{dictionary}
\keyword{sensata}
