% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictGenerator.R
\name{dictGenerator}
\alias{dictGenerator}
\title{Function to create sensata dictionary}
\usage{
dictGenerator(
  df,
  cols = c("identifier", "question", "type", "options", "numberOfOptions", "isOrdered",
    "isSorting", "maxResponses", "altOption"),
  expandOptions = TRUE,
  questionPrefix = "q_"
)
}
\arguments{
\item{df}{data downloaded from Mongo and loaded to R}

\item{cols}{columns to create on the dictionary.}

\item{expandOptions}{logical. If TRUE, the dictionary will have one row per answer option. if FALSE, it will be 1 row per question.}

\item{questionPrefix}{Character that identifies questions}
}
\value{
Dataframe dictionary useful for data cleaning
}
\description{
This function allows you to create a dictionary from Sensata's data exported from Mongo.
}
\examples{
dictGenerator(data = demoData, expandOptions = F)
}
\author{
Camilo Delvasto \email{camilo@sensata.io} & Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{diccionario}
\keyword{dictionary}
\keyword{sensata}
