% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dividethou_fun}
\alias{dividethou_fun}
\title{Miliseconds to seconds (or divide by a thousand)}
\usage{
dividethou_fun(x)
}
\arguments{
\item{x}{vector to convert from miliseconds to seconds (or divide by a thousand)}
}
\value{
vector of seconds or divided by a thousand
}
\description{
This function divides any vector by a thousand, we use it to convert seconds to milliseconds
}
\examples{
dividethou_fun(data$q1time)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{divide}
\keyword{seconds}
\keyword{segundos}
\keyword{sensata}
