% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{makeFactors}
\alias{makeFactors}
\title{Function to create factors from sensata data}
\usage{
makeFactors(
  df,
  dictionary,
  questionPrefix = "q_",
  skipQuestionString = "Saltar pregunta"
)
}
\arguments{
\item{df}{data downloaded from Mongo, cleaned with cleanData.R and scrubbed with scrubData.R}

\item{dictionary}{dictionary created using dictGenerator.R in expanded form}

\item{questionPrefix}{Character that identifies questions}

\item{skipQuestionString}{value that represents the skipped questions. By default "Saltar pregunta"}
}
\value{
Dataframe with questions in format that is analysis friendly.
}
\description{
This function allows you to create factors with the information from the dictionary to the variables in the microdata.
DO NOT USE for NPSor slider questions that have range different than 0-10.
}
\examples{
coronaTestData <-makeFactors(df = bogData3, dict = bogExpDic)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{dictionary}
\keyword{factors}
\keyword{microdata}
\keyword{sensata}
