% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{selectCols}
\alias{selectCols}
\title{Function to export sensata data for final customer}
\usage{
selectCols(
  df,
  dropGeo = T,
  dropParams = T,
  dropUserData = T,
  dropMetaData = T,
  dropTotalTime = T,
  dropQuestionTime = T
)
}
\arguments{
\item{df}{data downloaded from Mongo, cleaned with cleanData.R, scrubbed with scrubData.R and prepared with factorSensata.R.}

\item{dropTotalTime}{if TRUE it drops totalTimeMin variable.}

\item{dropQuestionTime}{if TRUE drops q_time variables}
}
\value{
Dataframe ready for client.
}
\description{
This function eliminates most metadata from the dataset. This is basically a wrapper for tidyverse::select()
}
\examples{
bogData1 <- bogData1 \%>\% scrubData(dup = F, ageVar = "EVCS2", ageVal = "Menos de 18 años", geoloc = T)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
