% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateFactors.R
\name{translateFactors}
\alias{translateFactors}
\title{Function to translate factors from sensata data two different languages}
\usage{
translateFactors(
  df,
  fromDictionary,
  toDictionary,
  questionPrefix = "q_",
  skipQuestionString = "Saltar pregunta"
)
}
\arguments{
\item{df}{data downloaded from Mongo, cleaned with cleanData.R and scrubbed with scrubData.R}

\item{fromDictionary}{dictionary created using dictGenerator.R in expanded form without screens in the original language}

\item{toDictionary}{dictionary created using dictGenerator.R in expanded form without screens in the target language}

\item{questionPrefix}{Character that identifies questions}

\item{skipQuestionString}{value that represents the skipped questions. By default "Saltar pregunta"}
}
\value{
Dataframe with questions in format that is analysis friendly.
}
\description{
This function allows you to create factors with the information from the dictionary to the variables in the microdata.
DO NOT USE for NPSor slider questions that have range different than 0-10.
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{dictionary}
\keyword{factors}
\keyword{microdata}
\keyword{sensata}
